moved reflection task to own widget
