+++ lib/screens/main_screen.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import '../task_box.dart';
+import '../widgets/task_box.dart';
 import '../buttons/success_btn.dart';
 import '../buttons/failure_btn.dart';
